pidcontroller works

+++ pidcontroller.h
@@ -7,54 +7,54 @@
 class EPID
 {
 		public:
-				EPID(double setpoint, double kp, double ki, double kd);
-				EPID(double kp, double ki, double kd);
-				void compute(double input, double& output, uint64_t deltatime = 0);
-				double compute(double input, uint64_t deltatime = 0);
+				EPID(float setpoint, float kp, float ki, float kd);
+				EPID(float kp, float ki, float kd);
+				void compute(float input, float& output, uint64_t deltatime = 0);
+				float compute(float input, uint64_t deltatime = 0);
 
 				// This may be funny.
-				void compute(double setpoint, double input, double& output);
+				void compute(float setpoint, float input, float& output);
 
-				void set_pid(double kp, double ki, double kd);
-				void set_point(double setpoint);
-				void set_output_limits(double min, double max);
+				void set_pid(float kp, float ki, float kd);
+				void set_point(float setpoint);
+				void set_output_limits(float min, float max);
 				void set_minsample_time(uint64_t sampletime);
-				void set_integral_limit(double limit);
+				void set_integral_limit(float limit);
 
-				double get_output();
+				float get_output();
 
 				// these could be commited for optimize.
-				double get_kp();
-				double get_ki();
-				double get_kd();
-				double get_setpoint();
-				double get_deltatime();
-				double get_minsample_time();
-				double get_proportional();
-				double get_integral();
-				double get_derivative();
+				float get_kp();
+				float get_ki();
+				float get_kd();
+				float get_setpoint();
+				float get_deltatime();
+				float get_minsample_time();
+				float get_proportional();
+				float get_integral();
+				float get_derivative();
 
 		private:
-				double KP;
-				double KI;
-				double KD;
+				float KP;
+				float KI;
+				float KD;
 
-				double input;
-				double setpoint;
-				double output;
+				float input;
+				float setpoint;
+				float output;
 
-				double error;
-				double pre_error;
-				double prepre_error;
+				float error;
+				float pre_error;
+				float prepre_error;
 
-				double proportional;
-				double integral;
-				double derivative;
+				float proportional;
+				float integral;
+				float derivative;
 
-				double integ_limit;
+				float integ_limit;
 				
-				double min_output;
-				double max_output;
+				float min_output;
+				float max_output;
 
 				uint64_t last_time;	//in us;
 				uint64_t delta_time;
